Add autodiff

+++ code/include/main_app/simulator/Constraint.hpp
@@ -3,10 +3,11 @@
 
 #include <functional>
 
+
 #include "Vector2d.hpp"
 
 namespace simulator {
-using Constraint = std::function<float(Vector2d position)>;
+using Constraint = std::function<double(Vector2d position)>;
 }
 
 #endif// CONSTRAINTBASEDSIMULATOR_SIMULATOR_CONSTRAINT_HPP
